requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytz>=2023.3
flask>=3.0.0
pillow>=10.0.0
anthropic>=0.18.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
redis>=5.0.0
supabase>=2.0.0


